ad_library_api/requirements.txt
appdirs==1.4.3
attrs==18.1.0
cached-property==1.4.3
cachetools==2.1.0
certifi==2018.1.18
chardet==3.0.4
defusedxml==0.5.0
facebook-sdk==3.1.0
google-auth==1.5.0
google-auth-oauthlib==0.2.0
googleads==12.2.0
idna==2.6
isodate==0.6.0
lxml==4.2.3
mixpanel==4.3.2
oauthlib==2.1.0
pyasn1==0.4.3
pyasn1-modules==0.2.2
Pygments==2.3.1
pytz==2018.5
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.0.0
requests-toolbelt==0.8.0
rsa==3.4.2
six==1.11.0
suds-jurko==0.6
twython==3.7.0
ujson==1.35
urllib3==1.24.2
xmltodict==0.11.0
zeep==3.0.0



